+++ extension/ts/objects/googleAuth.ts
