+++ extension/ts/objects/sheets.ts
@@ -1,4 +1,4 @@
-import Auth from "./auth";
+import {googleFetch} from "./googleFetch";
 
 const BASE_URL = "https://sheets.googleapis.com/v4/spreadsheets";
 
@@ -12,7 +12,6 @@ export interface GetSheetsDataResponse {
 	spreadsheetId: string;
 	valueRanges: ValueRange[];
 }
-
 interface UpdateSheetsDataResponse {
 	spreadsheetId: string;
 	updatedRange: string;
@@ -28,95 +27,62 @@ export interface AppendSheetsDataResponse {
 	updates: UpdateSheetsDataResponse;
 }
 
-const handleError = (res: Response) => {
-	if (res.status >= 400) {
-		Auth.clearAPICredentials();
-		alert(
-			"Something went wrong when trying to process that action. Please try again and ensure the inputted API key/ Client ID credentials are correct."
-		);
-	}
-};
-
-const readAllRowsFromSheet = async (spreadsheetId: string, ranges: string[]): Promise<GetSheetsDataResponse | null> => {
-	const credentials = Auth.getAPICredentials();
-
+const readAllRowsFromSheet = (spreadsheetId: string, ranges: string[]): Promise<GetSheetsDataResponse | null> => {
 	const queryParams = ranges.reduce((acc, range) => {
 		acc.append("ranges", range);
 		return acc;
 	}, new URLSearchParams({}));
-	queryParams.append("key", credentials.apiKey);
 
-	const res = await fetch(`${BASE_URL}/${spreadsheetId}/values:batchGet?${queryParams}`);
-
-	handleError(res);
-	return res.status == 200 ? res.json() : null;
+	return googleFetch(`${BASE_URL}/${spreadsheetId}/values:batchGet?${queryParams}`);
 };
 
-const appendRowToSheet = async (
+const appendRowToSheet = (
 	spreadsheetId: string,
 	range: string,
 	values: string[][]
 ): Promise<AppendSheetsDataResponse | null> => {
-	const credentials = Auth.getAPICredentials();
-	Auth.authorize();
-	const localStorageOAuth = Auth.getOAuthCredentials();
-	const token = localStorageOAuth.access_token;
-
 	const options: RequestInit = {
 		method: "POST",
 		headers: {
-			Authorization: `Bearer ${token}`,
 			Accept: "application/json",
 			"Content-Type": "application/json",
 		},
 		body: JSON.stringify({values}),
 	};
 
-	const res = await fetch(
+	return googleFetch(
 		`${BASE_URL}/${spreadsheetId}/values/${range}:append?${new URLSearchParams({
 			includeValuesInResponse: "true",
 			insertDataOption: "INSERT_ROWS",
 			valueInputOption: "RAW",
-			key: credentials.apiKey,
 		})}`,
 		options
 	);
-
-	handleError(res);
-	return res.status == 200 ? res.json() : null;
 };
 
-const updateRowInSheet = async (
+const updateRowInSheet = (
 	spreadsheetId: string,
 	range: string,
 	values: string[][]
 ): Promise<UpdateSheetsDataResponse | null> => {
-	const credentials = Auth.getAPICredentials();
-	Auth.authorize();
-	const localStorageOAuth = Auth.getOAuthCredentials();
-	const token = localStorageOAuth.access_token;
-
 	const options: RequestInit = {
 		method: "PUT",
 		headers: {
-			Authorization: `Bearer ${token}`,
 			Accept: "application/json",
 			"Content-Type": "application/json",
 		},
 		body: JSON.stringify({values}),
 	};
 
-	const res = await fetch(
+	return googleFetch(
 		`${BASE_URL}/${spreadsheetId}/values/${range}?${new URLSearchParams({
 			includeValuesInResponse: "true",
 			valueInputOption: "RAW",
-			key: credentials.apiKey,
 		})}`,
 		options
 	);
-
-	handleError(res);
-	return res.status == 200 ? res.json() : null;
 };
+
+//GoogleFetch parameterized
 
 export default {readAllRowsFromSheet, appendRowToSheet, updateRowInSheet};
